+++ App/Views/Settings/BellmanSettingsView.swift
@@ -38,7 +38,7 @@ struct BellmanSettingsView: View {
                 }
             },
             header: {
-                Label("Bellman Transceiver BT", systemImage: "hearingdevice.ear")
+                Label("Bellman Transceiver BT", systemImage: "hearingdevice.ear").dosHeader()
             }
         )
     }